refactor(tests): Split addSubtask tests
